build note item modal and header

+++ frontend/src/pages/Notes.jsx
@@ -2,6 +2,8 @@ import { useEffect, useState } from 'react'
 
 import Note from '../components/Note'
 import axios from 'axios'
+import BasicButton from '../components/BasicButton'
+import Modal from '../components/Modal'
 
 function Notes() {
   const API_URL = 'http://localhost:5000/api/notes/'
@@ -17,20 +19,22 @@ function Notes() {
   }, [])
 
   const onDelete = async (id) => {
-    try {
-      const res = await axios.delete(API_URL + id)
-      const newNotes = notes.filter((note) => note._id !== id)
-      setNotes(newNotes)
+    if (window.confirm('Delete note?')) {
+      try {
+        const res = await axios.delete(API_URL + id)
+        const newNotes = notes.filter((note) => note._id !== id)
+        setNotes(newNotes)
 
-      setIsDelete(true)
-    } catch (error) {
-      console.log(error)
+        setIsDelete(true)
+      } catch (error) {
+        console.log(error)
+      }
     }
   }
 
   return (
-    <div className='container mx-auto px-10'>
-      <div className='grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-6  md:justify-evenly '>
+    <div className='container mx-auto  px-10'>
+      <div className=' grid grid-cols-1 py-32 sm:grid-cols-2 lg:grid-cols-3 gap-6  md:justify-evenly '>
         {notes.map((note) => (
           <Note key={note._id} onDelete={onDelete} note={note} />
         ))}
